Add ability to delete columns from comparison table
This adds a remove button to the comparison table for each product that
is added. When clicked it uses a stimulus controller to find the index
of the nearest matching td and then locates the index of that td in the
row which is the index for the column which needs to be removed.

It then gets a list of all the rows of the table and iterates them. If
the row contains the dropdown class the iteration is skipped as the
dropdown rows should not be deleted. Then for any other rows it deletes
the cell at the matching index